Create hwintrinsic.cpp
