+++ src/jit/CMakeLists.txt
@@ -38,6 +38,7 @@ set( JIT_SOURCES
   gentree.cpp
   gschecks.cpp
   hashbv.cpp
+  hwintrinsic.cpp
   hostallocator.cpp
   importer.cpp
   inline.cpp
